fix: render dashboard avatar full bleed
